Print dictionary information

Print metadata for the dictionary name specified or for all available dictionaries (if name is unspecified)

Usage

dict_info(name = NA)

Arguments

name

string

Examples

dict_info()
#> Available dictionaries: 
#> 
#> Dictionary: dukecommunity2015
#> Country or context: US
#> Year: 2015
#> Description: 
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/
#> Citation: 
#> Notes: 
#> 
#> Dictionary: dukestudent2015
#> Country or context: US
#> Year: 2015
#> Description: 
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/
#> Citation: 
#> Notes: 
#> 
#> Dictionary: china1999
#> Country or context: China
#> Year: 1999
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 449 Identities, 300 Behaviors, 98 Emotions, 150 Traits, and 149 Settings were obtained with Attitude from about 380 undergraduate students at Fudan University in Shanghai, Peoples Republic of China, 1999-2000.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Smith, Herman, W. and Yi Cai. Mean Affective Ratings of 1,146 Concepts by Shanghai Undergraduates, 1999 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: egypt2015
#> Country or context: Egypt
#> Year: 2012-2014
#> Description: Data collected in Egypt between 2012 and 2014. n = 1716, identities = 397, behaviors = 368, modifiers = 233
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///egypt-dictionary-2015/
#> Citation: Hamid Latif, Lynn Smith-Lovin, Dawn T. Robinson, Bryan C. Cannon, Brent Curdy, Darys J Kriegel and Jonathan H. Morgan. 2016. "Mean Affective Ratings of Identities, Behaviors, and Modifiers by Residents of Cairo, Egypt in 2012-2014." University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/.
#> Notes: 
#> 
#> Dictionary: expressive2002
#> Country or context: US
#> Year: 2002
#> Description: Appendix of Lisa Slattery Rashotte, 2002, What does that smile mean? The meaning of nonverbal behaviors in social interaction. Social Psychology Quarterly 65: 92-102. EPA values are means across 230 females and 172 males; these cross-sex EPA profiles are given in both the "male" and the "female" positions.
#> Components: behavior
#> Genders: average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/expressive_acts.htm
#> Citation: Appendix of Lisa Slattery Rashotte, 2002, What does that smile mean? The meaning of nonverbal behaviors in social interaction. Social Psychology Quarterly 65: 92-102.
#> Notes: 
#> 
#> Dictionary: gaymensanfrancisco1980
#> Country or context: US
#> Year: 1980
#> Description: Data were gathered from ten San Francisco gay men in the 1980s by Professor Don Barrett, California State University, San Marcos. All data are from males. The 'male' dictionary present the mean EPA ratings of those with traditional sentiments about unsafe sex practices, feeling that those practices are more pleasurable than new practices. The 'female' dictionary shows the sentiments of those who think that safe-sex practices are more pleasurable.
#> Components: behavior
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/gay_sex.htm
#> Citation: 
#> Notes: All respondents are men. "Male" and "female" sets are coded this way for compatibility with Interact. 
#> 
#> Dictionary: germany1989
#> Country or context: Germany
#> Year: 1989
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 442 Identities, 295 Behaviors, and 67 Modifiers, selected for back-translatability with the 1978 U.S.A. dictionary were obtained with the Attitude program from 520 Mannheim students. Subjects were matched to the American undergraduate population by proportional inclusion of 12 and 13 grade youths at two German Studenten des Grundstudiums and Gymnasiasten, along with subjects from Mannheim University, which attracts students mainly from the Rhein-Neckar region in former West Germany.
#> Components: identity, behavior, modifier
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Schneider, Andreas. Mean Affective Ratings of 804 Concepts by West German Students  in 1989 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: germany2007
#> Country or context: Germany
#> Year: 2007
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 376 Identities, 393 Behaviors, 331 Modifiers. (Some of the words are classified as 19 settings in Interact.) Ratings were obtained with Surveyor from 1905 subjects (734 male and 1171 female) from all over Germany. The research was advertised as a 'study of language and emotion' in an extensive recruitment campaign including mailing lists from different universities, weblogs, newspaper reports and radio interviews. Most of the participants (N = 1029) were between 20 and 29 years of age, but the sample covered all ages, including N = 129 being younger than 20 and N = 92 older than 60 years. The data of 83 persons (4.4 %) were excluded from the analysis, as they had indicated German not being their mother tongue.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Schroder, Tobias. Mean Affective Ratings of 1100 Concepts by German Adults in 2007 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2013.
#> Notes: 
#> 
#> Dictionary: household1994
#> Country or context: US
#> Year: 1994
#> Description: This set of concepts - developed by Professor Amy Kroska at the University of Oklahoma, and others, for studying gender and household roles - was rated by 23 male and 46 female college students in 1994.
#> Components: identity, behavior
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/household.htm
#> Citation: 
#> Notes: 
#> 
#> Dictionary: indiana2003
#> Country or context: US
#> Year: 2003
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 500 Identities, 500 Behaviors, 300 Modifiers, and 200 Settings were collected at Indiana University, via the Internet using the Surveyor applet.  The 1027 respondents lived in the U.S.A. at age 16, and were about equally male and female.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Francis, Clare, and David R. Heise. Mean Affective Ratings of 1,500 Concepts by Indiana University Undergraduates in 2002-3 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: internationaldomesticrelations1981
#> Country or context: Unknown
#> Year: 1981
#> Description: 125 inter-state behaviors and 113 domestic behaviors. The behaviors were rated by professionals in international relations, and by individuals from the general population.
#> Components: behavior
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/international_relations.htm
#> Citation: Azar Edward E. and Steve Lerner 1981. The use of semantic dimensions in the scaling of international events. International Interactions 7: 361-378.
#> Notes: 
#> 
#> Dictionary: internet1998
#> Country or context: Internet
#> Year: 1998
#> Description: This set of concepts related to Internet culture was developed by Adam King, and rated by 2,431 respondents (56%  male and 44% female) in 1998.
#> Components: identity, behavior, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/InternetCulture.htm
#> Citation: King, Adam B. 2001. "Affective dimensions of internet culture." Social Science Computer Review 19:414-430.
#> Notes: 
#> 
#> Dictionary: japan1995
#> Country or context: Japan
#> Year: 1989-2002
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 403 Identities and 307 Behaviors, and a few Settings were obtained with the Attitude program from 323 Tohoku University students in 1989. In 1995 and 1996, 120 women students at Kyoritsu Women's, Japan Women's, and Teikyo Universities and 120 men students at Teikyo and Rikkyo Universities rated an additional 300 settings, 300 modifiers (mainly traits), 200 business identities, and 75 behaviors. Yoichi Murase (Rikkyo University) and Nozomu Matsubara (Tokyo University) provided access to students who rated 102 emotions, 70 behaviors and 55 identities in 2002 using Surveyor. Total numbers of entries in Interact lexicon are: 713 Identities, 455 Behaviors, 426 Modifiers, and 300 Settings. Number of male or female raters generally is about 30 for each concept.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Smith, Herman W., Takanori Matsuno, Shuuichirou Ike, and Michio Umino. Mean Affective Ratings of 1,894 Concepts by Japanese Undergraduates, 1989-2002 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: Data collection was in several stages between 1989 and 2002
#> 
#> Dictionary: morocco2015
#> Country or context: Morocco
#> Year: 2015
#> Description: Data collected in Morocco. n = 1546, identities = 397, behaviors = 368, modifiers = 233
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  No
#> Source: http://affectcontroltheory.org///morocco-dictionary-2015/
#> Citation: 
#> Notes: 
#> 
#> Dictionary: nc1978
#> Country or context: US
#> Year: 1978
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 721 Identities, 600 Behaviors, 440 Modifiers, and 345 Settings were obtained with paper questionnaires from 1,225 North Carolina undergraduates. (Ratings for some emotion words in this data set were obtained by Heise from Indiana University undergraduates in 1985.) Number of male or female raters generally is about 25 for each word. Funded by National Institute of Mental Health Grant 1-R01-MH29978-01-SSR.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Smith-Lovin, Lynn, and David R. Heise. Mean Affective Ratings of 2,106 Concepts by University of North Carolina Undergraduates in 1978 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: nireland1977
#> Country or context: North Ireland
#> Year: 1977
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 528 Identities and 498 Behaviors were obtained with paper questionnaires from 319 Belfast teenagers in Catholic high schools in 1977. Ratings of modifiers and settings were not obtained in the Belfast study. Up to 18 females and  14 males rated each concept.
#> Components: identity, behavior
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Willigan, Dennis, and David R. Heise. Mean Affective Ratings of 1,026 Concepts by Catholic High School Students in Belfast, Northern Ireland in 1977 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: ontario1980
#> Country or context: Canada
#> Year: 1980-1986
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Data on 843 Identities and  593 Behaviors were obtained from 5,534 Guelph, Ontario, undergraduates with paper questionnaires in 1980-3, and 495 Modifiers rated by 1,260 Guelph undergraduates were added in 1985-6. Funded by the Social Science and Humanities Research Council of Canada.
#> Components: identity, behavior, modifier
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: MacKinnon, Neil J. Mean Affective Ratings of 1,931 Concepts by Guelph University Undergraduates, Ontario, Canada, in 1980-6 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: ontario2001
#> Country or context: Canada
#> Year: 2001-2003
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Data on 993 Identities, 601 Behaviors, and 500 Modifiers, and 200 Settings were gathered with the Attitude program from Guelph, Ontario, undergraduates in 2001-2. Data on settings were gathered with the Surveyor program at Guelph in 2003. Funded by the Social Science and Humanities Research Council of Canada.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: MacKinnon, Neil J. Mean Affective Ratings of 2,294 Concepts by Guelph University Undergraduates, Ontario, Canada, in 2001-3 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: politics2003
#> Country or context: US
#> Year: 2003
#> Description: This set of concepts - developed by Kyle Irwin at the University of Missouri-St. Louis - was rated by 47 male and 74 female college students in 2003.
#> Components: identity, behavior
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/politics.htm
#> Citation: 
#> Notes: 
#> 
#> Dictionary: texas1998
#> Country or context: US
#> Year: 1998
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 443 Identities, 278 Behaviors, 65 Modifiers, and 1 Setting were collected at Texas Tech University with the Attitude program. The 482 respondents were nearly equally male and female.
#> Components: identity, behavior, modifier
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Schneider, Andreas. Mean Affective Ratings of 787 Concepts by Texas Tech University Undergraduates in 1998 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: 
#> 
#> Dictionary: uga2015
#> Country or context: US
#> Year: 2015
#> Description: Sentiment dictionary data collected at the University of Georgia. n = 1368, identities = 930, behaviors = 814, modifiers = 660
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-georgia-dictionary-2015/
#> Citation: Robinson, Dawn T., Lynn Smith-Lovin, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. 'Mean Affective Ratings of 932 Identities, 810 Behaviors, and 660 Modifiers by University of Georgia Undergraduates in 2012-2014.'. University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/.
#> Notes: 
#> 
#> Dictionary: usfullsurveyor2015
#> Country or context: US
#> Year: 2015
#> Description: Combined sentiment dictionary data collected at the University of Georgia, Duke University, and in the Durham, NC community using Surveyor. n = 1742, identities = 929, behaviors = 814, modifiers = 660
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/
#> Citation: Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. "Mean Affective Ratings of 929 Identities, 814 Behaviors, and 660 Modifiers by University of Georgia and Duke University Undergraduates and by Community Members in Durham, NC, in 2012-2014."" University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/.
#> Notes: A combination of the 2015 Duke student, 2015 UGA student, and 2015 Duke community dictionaries
#> 
#> Dictionary: usmturk2015
#> Country or context: US
#> Year: 2015
#> Description: Collected from Amazon Mechanical Turk workers in 2015. Identities = 968, behaviors = 853, modifiers = 660.
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-online-dictionary-2015/
#> Citation: Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Brent H. Curdy, and Jonathan H. Morgan. 2019. 'Mean Affective Ratings of 968 Identities, 853 Behaviors, and 660 Modifiers by Amazon Mechanical Turk Workers in 2015.' University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/
#> Notes: 
#> 
#> Dictionary: usstudent2015
#> Country or context: US
#> Year: 2015
#> Description: Combined sentiment dictionary data collected at the University of Georgia and Duke University using Surveyor. n = 1584, identities = 929, behaviors = 814, modifiers = 660
#> Components: identity, behavior, modifier
#> Genders: average
#> Stats: mean, sd, cov
#> Individual data available?  Yes
#> Source: http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/
#> Citation: Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. 'Mean Affective Ratings of 929 Identities, 814 Behaviors, and 660 Modifiers by University of Georgia and Duke University Undergraduates in 2012-2014.' University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/.
#> Notes: A combination of the 2015 Duke student and 2015 UGA student dictionaries
#> 
dict_info("china1999")
#> Dictionary: china1999
#> Country or context: China
#> Year: 1999
#> Description: From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 449 Identities, 300 Behaviors, 98 Emotions, 150 Traits, and 149 Settings were obtained with Attitude from about 380 undergraduate students at Fudan University in Shanghai, Peoples Republic of China, 1999-2000.
#> Components: identity, behavior, modifier, setting
#> Genders: male, female, average
#> Stats: mean
#> Individual data available?  No
#> Source: Interact 2.1 beta (May 2021)
#> Citation: Smith, Herman, W. and Yi Cai. Mean Affective Ratings of 1,146 Concepts by Shanghai Undergraduates, 1999 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006.
#> Notes: