epa_summary

Calculates the mean, standard deviation, and variance-covariance matrix for each term in an individual-level data set. This is useful when a user wants summary EPA information for a subset of respondents, for example, when comparing cultural meaning across groups. In this case, a user would first create the desired individual data subsets using the epa_subset() function, then pass the resulting data frames to this function to calculate relevant summary statistics.

Usage

epa_summary(data)

Arguments

data

individual level data frame with columns for term, component, E, P, and A.

Value

a summary dataset with one row per term. Includes the evaluation, potency, and activity mean, standard deviation, and variance-covariance matrix entries for each term/component combination. Values are rounded to the nearest .01.

Examples

epa_summary(dplyr::filter(epa_subset(datatype = "individual", dataset = "usfullsurveyor2015"),
    gender == "Male"))
#> # A tibble: 2,403 × 20
#>    term       component   n_E   n_P   n_A mean_E mean_P mean_A  sd_E  sd_P  sd_A
#>    <chr>      <chr>     <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 abandon    behavior     31    31    31  -3.47   0.37  -0.22  1     3.14  2.46
#>  2 abandoned  modifier     22    22    22  -3.15   0.42  -1.19  1.14  2.92  1.92
#>  3 abduct     behavior     48    48    48  -3.58   1.79   1.65  1.26  2.23  2.22
#>  4 abet       behavior     18    18    18  -0.16   0.22   0.08  1.01  1.24  0.92
#>  5 abhor      behavior      6     6     6  -2.75   0.76  -0.28  1.36  2.81  2.39
#>  6 able_bodi… modifier     31    31    31   3      2.81   1.22  1.09  1.34  1.81
#>  7 abort      behavior     26    26    26  -2.18   0.72  -0.21  1.83  2.94  2.61
#>  8 abortioni… identity     39    39    39  -0.86   1.11   0.03  1.98  2.43  1.98
#>  9 absent_mi… modifier     34    34    34  -1.72  -1.55  -1.47  1.5   1.78  1.92
#> 10 abuse      behavior     32    32    32  -3.64  -0.54   1.73  1.22  3.12  1.85
#> # ℹ 2,393 more rows
#> # ℹ 9 more variables: cov_EE <dbl>, cov_EP <dbl>, cov_EA <dbl>, cov_PE <dbl>,
#> #   cov_PP <dbl>, cov_PA <dbl>, cov_AE <dbl>, cov_AP <dbl>, cov_AA <dbl>