EPA summary statistics, all datasets

This data frame contains summary statistics for EPA ratings collected in many different data collection efforts. The function epa_subset() provides functionality for easily filtering it in desired ways, and I recommend using that function rather than working with this data frame directly.

Usage

epa_summary_statistics

Format

A data frame with 72951 rows and 25 variables:

term

dictionary term

component

the term's type

dataset

what data collection the statistics originate from

context

the country or other context data were collected in

year

the approximate year of data collection

group

the subset of respondents summary statistics are calculated for. Groups are usually genders–male, female, all–but some data sets use other grouping variables. See data set documentation.

instcodes

institution codes

E

evaluation mean value

P

potency mean value

A

activity mean value

n_E

number of respondents who gave valid E ratings

n_P

number of respondents who gave valid P ratings

n_A

number of respondents who gave valid A ratings

sd_E

evaluation standard deviation

sd_P

potency standard deviation

sd_A

activity standard deviation

cov_EE

covariance

cov_EP

covariance

cov_EA

covariance

cov_PE

covariance

cov_PP

covariance

cov_PA

covariance

cov_AE

covariance

cov_AP

covariance

cov_AA

covariance

...

Source

http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/

Details

For more information on data collections, call dict_info().