All equation dataframes

This nested data frame provides information and coefficients for ACT equations. Coefficient values are reported as given in Interact. If group is listed as all, this indicates that for the given study and subset, the values provided in Interact for male and female are identical. To retrieve coefficients for particular equation sets, I recommend using the get_eqn() function rather than working with this data frame directly.

Usage

equations

Format

A nested data frame with 78 rows and 4 columns

key

equation set name

equation_type

type of equation

group

equation gender (all, female, male)

df

dataframe with equation coefficients

...

Source

Interact (May 2021)

Details

For more information about available equation files, call eqn_info()