Extract single equation data frame

Extract single equation data frame

Usage

get_eqn(key, equation_type = "impressionabo", group = "all")

Arguments

key

the name of the equation set, Call eqn_info() or see package readme for options.

equation_type

the type of the equation. Options: emotionid, impressionabo, impressionabos, selfdir, traitid

group

respondent group (currently, this refers exclusively to gender groups, though in principle equations could be calculated for any subset of respondents). All, female, or male.

Value

equation dataframe

Examples

get_eqn("us2010")
#>            V1    V2    V3    V4    V5    V6    V7    V8    V9   V10
#> 1  Z000000000 -0.16 -0.06 -0.05 -0.26  0.09  0.21 -0.18  0.00 -0.14
#> 2  Z100000000  0.62  0.00  0.00  0.52  0.00  0.00  0.00  0.00  0.00
#> 3  Z010000000  0.00  0.57  0.00  0.00  0.53  0.00  0.00  0.00  0.00
#> 4  Z001000000  0.00  0.00  0.26  0.00  0.00  0.15  0.00  0.00  0.00
#> 5  Z000100000  0.46 -0.29 -0.18  0.50 -0.21 -0.12  0.26  0.51  0.33
#> 6  Z000010000  0.00  0.62  0.12  0.00  0.64  0.00  0.00 -0.57 -0.41
#> 7  Z000001000  0.00  0.00  0.86  0.00  0.00  0.91  0.00  0.00  0.31
#> 8  Z000000100  0.00  0.00  0.00  0.00  0.00  0.00  0.91  0.00  0.00
#> 9  Z000000010  0.00 -0.20  0.00  0.00  0.00  0.00  0.00  0.47  0.00
#> 10 Z000000001  0.00  0.00  0.00  0.00  0.00  0.09  0.00  0.00  0.54
#> 11 Z100100000  0.00  0.00  0.00  0.00  0.00  0.00  0.16  0.00  0.54
#> 12 Z000100100  0.29  0.00  0.00  0.32  0.00  0.00  0.11  0.00  0.00
#> 13 Z000010100 -0.22  0.00  0.00 -0.27  0.00  0.00 -0.15  0.00  0.00
#> 14 Z000100001 -0.09  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 15 Z001001100  0.00  0.00  0.00  0.00  0.00  0.00  0.00 -0.10  0.00
get_eqn("nc1978", equation_type = "impressionabos", group = "male")
#>               V1    V2    V3    V4    V5    V6    V7    V8    V9   V10   V11
#> 1  Z000000000000 -0.25 -0.14  0.08 -0.13  0.06  0.00 -0.10 -0.43 -0.03 -0.35
#> 2  Z100000000000  0.35  0.00  0.06  0.10  0.00  0.00  0.00  0.00  0.00  0.00
#> 3  Z010000000000  0.00  0.46 -0.05  0.00  0.13 -0.06  0.00  0.00  0.00  0.00
#> 4  Z001000000000  0.00  0.08  0.65  0.00  0.00  0.27  0.00  0.00  0.00  0.00
#> 5  Z000100000000  0.52 -0.08 -0.08  0.56 -0.12 -0.06  0.17  0.19  0.03  0.11
#> 6  Z000010000000 -0.05  0.46  0.10 -0.06  0.68  0.12  0.00 -0.12  0.00  0.00
#> 7  Z000001000000 -0.09  0.00  0.27 -0.12  0.00  0.61  0.00  0.05  0.05  0.00
#> 8  Z000000100000  0.00  0.00  0.00  0.00  0.00  0.00  0.49 -0.09  0.00  0.00
#> 9  Z000000010000  0.00  0.00  0.00 -0.14  0.00  0.00 -0.14  0.38 -0.05  0.00
#> 10 Z000000001000  0.12  0.00  0.00  0.00  0.00  0.04  0.00  0.08  0.48  0.00
#> 11 Z000000000100  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.08  0.00  0.60
#> 12 Z000000000010  0.00  0.00  0.00  0.00  0.09  0.00  0.00  0.00  0.00  0.00
#> 13 Z000000000001  0.07  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00 -0.06
#> 14 Z100100000000  0.05  0.00  0.00  0.01  0.00  0.00  0.03  0.01  0.00  0.00
#> 15 Z100010000000 -0.04  0.00  0.00  0.00  0.00  0.00  0.00  0.03  0.00  0.00
#> 16 Z100000100000  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00 -0.02  0.00
#> 17 Z010100000000  0.00  0.05  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 18 Z010010000000  0.00 -0.07  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 19 Z010001000000  0.00  0.00  0.00  0.00  0.00  0.00  0.00 -0.15  0.00  0.00
#> 20 Z010000100000  0.08  0.00  0.00  0.09  0.00  0.00  0.00  0.00  0.00  0.00
#> 21 Z010000010000  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00 -0.03  0.00
#> 22 Z010000001000  0.00  0.00  0.00  0.00  0.03  0.00  0.00  0.00  0.00  0.00
#> 23 Z001010000000  0.00  0.00 -0.06  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 24 Z001001000000  0.00 -0.03 -0.05  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 25 Z000100100000  0.12  0.02  0.00  0.11  0.02  0.00  0.04  0.03  0.01  0.03
#> 26 Z000100010000 -0.06 -0.02  0.00 -0.05  0.00  0.00  0.00  0.02  0.00  0.00
#> 27 Z000010100000 -0.05  0.00  0.00  0.00  0.00  0.00 -0.02  0.00  0.00  0.00
#> 28 Z000010010000 -0.01  0.00  0.00  0.04  0.00  0.00  0.00  0.00  0.00  0.00
#> 29 Z000010001000  0.00  0.03  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#> 30 Z000001010000  0.00  0.00  0.00  0.03  0.00  0.00  0.00  0.00  0.00  0.00
#> 31 Z100100100000  0.02  0.01  0.00  0.02  0.01  0.00  0.01  0.00  0.00  0.00
#> 32 Z100010010000  0.03  0.00  0.00  0.02  0.00  0.00  0.00  0.00  0.00  0.00
#> 33 Z010010010000  0.00  0.00  0.00 -0.02  0.00  0.02  0.00 -0.02  0.00  0.00
#> 34 Z010010001000  0.00  0.00 -0.02  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#>      V12   V13
#> 1  -0.06  0.09
#> 2   0.00  0.00
#> 3   0.00  0.00
#> 4   0.00  0.07
#> 5   0.00  0.00
#> 6   0.10  0.00
#> 7   0.00  0.00
#> 8   0.00  0.00
#> 9   0.00  0.00
#> 10  0.00  0.00
#> 11 -0.14 -0.07
#> 12  0.66  0.00
#> 13  0.00  0.74
#> 14  0.00 -0.03
#> 15  0.00  0.00
#> 16  0.00  0.00
#> 17  0.00  0.00
#> 18  0.00  0.00
#> 19  0.00  0.00
#> 20  0.00  0.00
#> 21  0.00  0.00
#> 22  0.00  0.00
#> 23  0.00  0.00
#> 24  0.00  0.00
#> 25  0.00  0.00
#> 26  0.00  0.00
#> 27  0.00  0.00
#> 28  0.00  0.00
#> 29  0.06  0.00
#> 30  0.00  0.00
#> 31  0.00  0.00
#> 32  0.00  0.00
#> 33  0.00  0.00
#> 34  0.00  0.00