Terms

Terms of all types contained in all available dictionaries with indicators for which dictionary contains them. Note that terms have been standardized between dictionaries to facilitate comparison, so terms here may differ from those in the original source data files in substantively unimportant ways (capitalization, spacing, spelling, etc). There is one column per dictionary that includes this type of term. Whether or not a specific dictionary includes a particular term is indicated by 0/1 indicators.

Usage

term_table

Format

A data frame with 5608 rows and 27 variables:

term

concept

component

component this term represents (identity, behavior, modifier, or setting)

artifactmods2022

dictionary

calcuttaall2017

dictionary

calcuttasubset2017

dictionary

china1999

dictionary

dukecommunity2015

dictionary

dukestudent2015

dictionary

egypt2015

dictionary

employeeorg2022

dictionary

expressive2002

dictionary

gaymensanfrancisco1980

dictionary

generaltech2020

dictionary

germany1989

dictionary

germany2007

dictionary

groups2017

dictionary

groups2019

dictionary

household1994

dictionary

humanvalues2022

dictionary

indiana2003

dictionary

internationaldomesticrelations1981

dictionary

internet1998

dictionary

japan1995

dictionary

morocco2015

dictionary

mostafaviestimates2022

dictionary

nc1978

dictionary

nireland1977

dictionary

nounphrasegrammar2019

dictionary

occs2019

dictionary

occs2020

dictionary

occs2021

dictionary

ontario1980

dictionary

ontario2001

dictionary

politics2003

dictionary

products2022

dictionary

techvshuman2021

dictionary

texas1998

dictionary

ugatech2008

dictionary

uga2015

dictionary

usfullsurveyor2015

dictionary

usmturk2015

dictionary

usstudent2015

dictionary

...