Get dictionary or equation object

Extracts metadata object for a single dictionary of a given name. Its main purpose is to allow for testing whether particular dictionaries have particular elements.

Usage

this_dict(name)

Arguments

name

string

Value

dictionary object

Examples

this_dict("uga2015")
#> An object of class "dictionary"
#> Slot "key":
#> [1] "uga2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-georgia-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Sentiment dictionary data collected at the University of Georgia. n = 1368, identities = 930, behaviors = 814, modifiers = 660"
#> 
#> Slot "citation":
#> [1] "Robinson, Dawn T., Lynn Smith-Lovin, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. 'Mean Affective Ratings of 932 Identities, 810 Behaviors, and 660 Modifiers by University of Georgia Undergraduates in 2012-2014.'. University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/."
#> 
#> Slot "notes":
#> [1] ""
#>