Create institution code strings from logicals indicating category membership

This function returns institution codes in a string format that is properly formatted for import to Interact and consistent with typical ACT data practices. Relevant arguments depend on the component. All function parameters are logical, and all defaults are FALSE, meaning that to create an institution code, users need only set the categories that apply to TRUE. Arguments that do not apply to the given component will be ignored. See Heise's 2014 PDF manual for Interact or Heise's 2007 book Expressive Order for more details on these categories.

Usage

create_instcode(
  component,
  male = F,
  female = F,
  overt = F,
  surmised = F,
  place = F,
  time = F,
  lay = F,
  business = F,
  law = F,
  politics = F,
  academe = F,
  medicine = F,
  religion = F,
  family = F,
  sexual = F,
  monadic = F,
  group = F,
  corporal = F,
  adjective = F,
  adverb = F,
  emotion = F,
  trait = F,
  status = F,
  feature = F,
  emotion_spiral = F
)

Arguments

component

whether the term is an identity, behavior, component, or setting

male, female

logical. What genders terms can typically be applied to (identities only)

overt, surmised

logical. Whether labeling behaviors requires interpretation or insight on the part of the observer (behaviors only).

place, time

logical. Type of setting (settings only).

lay, business, law, politics, academe, medicine, religion, family, sexual

logical. Societal institutions that terms may belong to. Institutions, behaviors, and settings only.

monadic, group, corporal

logical. How a term requires or implicates others. Identities, behaviors, and settings only.

adjective, adverb

logical. Part of speech for modifiers.

emotion, trait, status, feature, emotion_spiral

logical. Categories for modifiers.

Value

character string institution code

Details

Identities: male, female, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal

Behaviors: overt, surmised, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal

Modifiers: adjective, adverb, emotion, trait, status, feature, emotion_spiral

Settings: place, time, lay, business, law, politics, academe, medicine, religion, family, sexual, monadic, group, corporal

Examples

create_instcode(component = "identity", male = TRUE, female = TRUE, lay = TRUE)
#> [1] "11 100000000 000"
create_instcode(component = "behavior", overt = TRUE, business = TRUE, group = TRUE)
#> [1] "10 010000000 010"
create_instcode(component = "modifier", adjective = TRUE, emotion = TRUE)
#> [1] "10 100000000 000"