EPA dictionary search and subset

Returns a subset of the EPA summary or individual data that fulfills the given parameters. Filtering can be done by term, data set, component (identity, behavior, modifier, setting), type of data (summary or individual), statistics (mean, standard deviation, covariance), institutions the term belongs to, and gender of raters.

Usage

epa_subset(
  expr = ".*",
  exactmatch = FALSE,
  dataset = "all",
  component = "all",
  datatype = "summary",
  gender = "all",
  stat = "all",
  stat_na_exclude = TRUE,
  instcodes = TRUE,
  institutions = "all",
  drop.na.instcodes = FALSE
)

Arguments

expr

A term, regular expression, or list of terms or regexs to search. If a list is provided, entries will be treated as separated by "or", so all terms matching one or more of the entries will be returned. Default matches all terms.

exactmatch

Logical indicating whether the function should return only exact matches to the expression provided. If FALSE (default), all terms containing the expression are returned.

dataset

The key of the data set (or list of multiple) to search in. Default is "all. Call dict_info() to see available data sets.

component

The component of the dictionary to use (identity, behavior, modifier, setting). Default is "all."

datatype

Whether to retrieve summary or individual data. Default is summary.

gender

The gender of the dictionary to use (male, female, average). Default is "all." Ignored when datatype is individual.

stat

The statistics to include in the subset that is returned. Default is all, options are mean, sd (standard deviation), cov (covariance), and n (number of raters). Terms that do not contain values for the required statistic will be excluded from the results. Ignored if datatype is individual.

stat_na_exclude

Ignored if stat is not specified of datatype is individual. A logical indicating whether to exclude entries with NA values for any of the required statistics. Default is TRUE.

instcodes

Logical. Whether to include the institution codes in the output. Default is TRUE.

institutions

Character list. Institutions to include (defaults to all)

drop.na.instcodes

Logical. When filtering by institution, whether or not to keep terms which have no institution code.

Value

a dataset containing the entries that match the given parameters or FALSE if no matches are found.

Examples

epa_subset("teacher")
#> # A tibble: 177 × 25
#>    term     compo…¹ dataset context year  gender instc…²     E     P     A   n_E
#>    <chr>    <chr>   <chr>   <chr>   <chr> <chr>  <chr>   <dbl> <dbl> <dbl> <dbl>
#>  1 school_… identi… calcut… India   2017  male   NA       2.02  1.47  0.83    20
#>  2 school_… identi… calcut… India   2017  female NA       1.88  1.14  1.01    20
#>  3 school_… identi… calcut… India   2017  avera… NA       1.95  1.3   0.92    40
#>  4 school_… identi… calcut… India   2017  male   NA       2.15  1.72  1.95    20
#>  5 school_… identi… calcut… India   2017  female NA       2.27  1.83  1.47    20
#>  6 school_… identi… calcut… India   2017  avera… NA       2.21  1.77  1.71    40
#>  7 schoolt… identi… calcut… India   2017  male   NA       2.09  1.47  1.62    20
#>  8 schoolt… identi… calcut… India   2017  female NA       1.68  1.39  0.5     20
#>  9 schoolt… identi… calcut… India   2017  avera… NA       1.89  1.43  1.06    40
#> 10 teacher  identi… calcut… India   2017  male   11 000…  2.2   1.99  1.91    20
#> # … with 167 more rows, 14 more variables: n_P <dbl>, n_A <dbl>, sd_E <dbl>,
#> #   sd_P <dbl>, sd_A <dbl>, cov_EE <dbl>, cov_EP <dbl>, cov_EA <dbl>,
#> #   cov_PE <dbl>, cov_PP <dbl>, cov_PA <dbl>, cov_AE <dbl>, cov_AP <dbl>,
#> #   cov_AA <dbl>, and abbreviated variable names ¹​component, ²​instcodes
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
epa_subset(dataset = "politics2003")
#> # A tibble: 216 × 25
#>    term   compo…¹ dataset context year  gender instc…²      E     P      A   n_E
#>    <chr>  <chr>   <chr>   <chr>   <chr> <chr>  <chr>    <dbl> <dbl>  <dbl> <dbl>
#>  1 alder… identi… politi… US      2003  avera… NA       0.735 0.905  0.335    NA
#>  2 alder… identi… politi… US      2003  male   NA       0.82  0.96   0.15     NA
#>  3 alder… identi… politi… US      2003  female NA       0.65  0.85   0.52     NA
#>  4 assem… identi… politi… US      2003  avera… NA       1.38  1.24   1.09     NA
#>  5 assem… identi… politi… US      2003  male   NA       1.51  1.05   1.06     NA
#>  6 assem… identi… politi… US      2003  female NA       1.25  1.43   1.12     NA
#>  7 burea… identi… politi… US      2003  avera… NA      -0.085 1.23  -0.115    NA
#>  8 burea… identi… politi… US      2003  male   NA      -0.43  0.91  -0.86     NA
#>  9 burea… identi… politi… US      2003  female NA       0.26  1.54   0.63     NA
#> 10 burea… identi… politi… US      2003  avera… 11 011…  0.415 0.665  0.075    NA
#> # … with 206 more rows, 14 more variables: n_P <dbl>, n_A <dbl>, sd_E <dbl>,
#> #   sd_P <dbl>, sd_A <dbl>, cov_EE <dbl>, cov_EP <dbl>, cov_EA <dbl>,
#> #   cov_PE <dbl>, cov_PP <dbl>, cov_PA <dbl>, cov_AE <dbl>, cov_AP <dbl>,
#> #   cov_AA <dbl>, and abbreviated variable names ¹​component, ²​instcodes
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
epa_subset(expr = ".*woman", component = "identity", gender = c("male", "female"),
    institutions = c("lay", "business"))
#> # A tibble: 72 × 25
#>    term     compo…¹ dataset context year  gender instc…²     E     P     A   n_E
#>    <chr>    <chr>   <chr>   <chr>   <chr> <chr>  <chr>   <dbl> <dbl> <dbl> <dbl>
#>  1 battere… identi… calcut… India   2017  male   01 101… -0.58 -0.22  0.33    12
#>  2 battere… identi… calcut… India   2017  female 01 101…  0     0.43 -0.61    16
#>  3 pregnan… identi… calcut… India   2017  male   NA       0.46  0.27  1.02    16
#>  4 pregnan… identi… calcut… India   2017  female NA       1.97  1.23  0.56    19
#>  5 woman    identi… calcut… India   2017  male   01 100…  1.24  0.58 -0.03    15
#>  6 woman    identi… calcut… India   2017  female 01 100…  1.2   0.55  0.85    14
#>  7 woman_w… identi… calcut… India   2017  male   NA       0.55  0.34  0.11    20
#>  8 woman_w… identi… calcut… India   2017  female NA       1.04  0.57  0       20
#>  9 battere… identi… calcut… India   2017  male   01 101… -0.58 -1.37 -2.02    12
#> 10 battere… identi… calcut… India   2017  female 01 101…  0     0.13  0.01    16
#> # … with 62 more rows, 14 more variables: n_P <dbl>, n_A <dbl>, sd_E <dbl>,
#> #   sd_P <dbl>, sd_A <dbl>, cov_EE <dbl>, cov_EP <dbl>, cov_EA <dbl>,
#> #   cov_PE <dbl>, cov_PP <dbl>, cov_PA <dbl>, cov_AE <dbl>, cov_AP <dbl>,
#> #   cov_AA <dbl>, and abbreviated variable names ¹​component, ²​instcodes
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
epa_subset(dataset = "morocco2015", stat = "cov", stat_na_exclude = FALSE)
#> # A tibble: 1,448 × 16
#>    term        compo…¹ dataset context year  gender instc…² cov_EE cov_EP cov_EA
#>    <chr>       <chr>   <chr>   <chr>   <chr> <chr>  <chr>    <dbl>  <dbl>  <dbl>
#>  1 abandon     behavi… morocc… Morocco 2015  avera… 01 111…   4.43   1.76  -1.6 
#>  2 abortionist identi… morocc… Morocco 2015  avera… 11 000…   5.5    1.66  -1.39
#>  3 abuse       behavi… morocc… Morocco 2015  avera… 10 101…   4.12   0.37  -1.69
#>  4 abusive     modifi… morocc… Morocco 2015  avera… 10 010…   5.98   3.34  -1.1 
#>  5 accommodat… modifi… morocc… Morocco 2015  avera… 10 010…   4.38   2.8   -0.87
#>  6 accuse      behavi… morocc… Morocco 2015  avera… 10 111…   6      1.66  -1.15
#>  7 address     behavi… morocc… Morocco 2015  avera… 10 111…   3.61   1.43  -1.68
#>  8 admonish    behavi… morocc… Morocco 2015  avera… 10 011…   4.4    1.23  -1.88
#>  9 adolescent  identi… morocc… Morocco 2015  avera… 11 100…   3.03   1.23  -0.19
#> 10 adult       identi… morocc… Morocco 2015  avera… 11 100…   4.21   2.43   0.3 
#> # … with 1,438 more rows, 6 more variables: cov_PE <dbl>, cov_PP <dbl>,
#> #   cov_PA <dbl>, cov_AE <dbl>, cov_AP <dbl>, cov_AA <dbl>, and abbreviated
#> #   variable names ¹​component, ²​instcodes
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
epa_subset(dataset = "usmturk2015", datatype = "individual")
#> # A tibble: 264,844 × 16
#>    dataset   context year  userid gender age   race1 race2 term  compo…¹ instc…²
#>    <chr>     <chr>   <chr> <chr>  <chr>  <chr> <chr> <chr> <chr> <chr>   <chr>  
#>  1 usmturk2… US      2015  MTurk1 Female 51    1     6     tele… identi… NA     
#>  2 usmturk2… US      2015  MTurk1 Female 51    1     6     bewi… behavi… 01 000…
#>  3 usmturk2… US      2015  MTurk1 Female 51    1     6     coal… identi… 10 010…
#>  4 usmturk2… US      2015  MTurk1 Female 51    1     6     unde… behavi… 01 111…
#>  5 usmturk2… US      2015  MTurk1 Female 51    1     6     appl… behavi… 10 111…
#>  6 usmturk2… US      2015  MTurk1 Female 51    1     6     gran… identi… 11 000…
#>  7 usmturk2… US      2015  MTurk1 Female 51    1     6     love  behavi… 01 100…
#>  8 usmturk2… US      2015  MTurk1 Female 51    1     6     barr… behavi… 10 000…
#>  9 usmturk2… US      2015  MTurk1 Female 51    1     6     humb… behavi… 01 111…
#> 10 usmturk2… US      2015  MTurk1 Female 51    1     6     land… identi… 10 010…
#> # … with 264,834 more rows, 5 more variables: E <chr>, P <chr>, A <chr>,
#> #   race <chr>, eth <chr>, and abbreviated variable names ¹​component,
#> #   ²​instcodes
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names