Get dictionary information

Return metadata of available dictionaries as a list of dictionary objects

Usage

get_dicts()

Value

list of dictionary objects

Examples

get_dicts()
#> [[1]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "calcuttaall2017"
#> 
#> Slot "context":
#> [1] "India"
#> 
#> Slot "year":
#> [1] "2017"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting"  "artifact"
#> 
#> Slot "stats":
#> [1] "mean" "sd"  
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "https://link.springer.com/article/10.3758/s13428-016-0704-6"
#> 
#> Slot "description":
#> [1] "Data collection from 20 male and 20 female respondents in Calcutta"
#> 
#> Slot "citation":
#> [1] "Mukherjee, S., & Heise, D. R. (2017). Affective meanings of 1,469 Bengali concepts. Behavior Research Methods, 49(1), 184-197."
#> 
#> Slot "notes":
#> [1] "Results from all respondents regardless of scale use (see paper for details)"
#> 
#> 
#> [[2]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "calcuttasubset2017"
#> 
#> Slot "context":
#> [1] "India"
#> 
#> Slot "year":
#> [1] "2017"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting"  "artifact"
#> 
#> Slot "stats":
#> [1] "mean" "sd"  
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "https://link.springer.com/article/10.3758/s13428-016-0704-6"
#> 
#> Slot "description":
#> [1] "Data collection from 20 male and 20 female respondents in Calcutta"
#> 
#> Slot "citation":
#> [1] "Mukherjee, S., & Heise, D. R. (2017). Affective meanings of 1,469 Bengali concepts. Behavior Research Methods, 49(1), 184-197."
#> 
#> Slot "notes":
#> [1] "Results from only those respondents who use scales as expected (see paper for details)"
#> 
#> 
#> [[3]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "china1999"
#> 
#> Slot "context":
#> [1] "China"
#> 
#> Slot "year":
#> [1] "1999"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 449 Identities, 300 Behaviors, 98 Emotions, 150 Traits, and 149 Settings were obtained with Attitude from about 380 undergraduate students at Fudan University in Shanghai, Peoples Republic of China, 1999-2000."
#> 
#> Slot "citation":
#> [1] "Smith, Herman, W. and Yi Cai. Mean Affective Ratings of 1,146 Concepts by Shanghai Undergraduates, 1999 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[4]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "dukecommunity2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[5]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "dukestudent2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[6]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "egypt2015"
#> 
#> Slot "context":
#> [1] "Egypt"
#> 
#> Slot "year":
#> [1] "2012-2014"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///egypt-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Data collected in Egypt between 2012 and 2014. n = 1716, identities = 397, behaviors = 368, modifiers = 233"
#> 
#> Slot "citation":
#> [1] "Hamid Latif, Lynn Smith-Lovin, Dawn T. Robinson, Bryan C. Cannon, Brent Curdy, Darys J Kriegel and Jonathan H. Morgan. 2016. \"Mean Affective Ratings of Identities, Behaviors, and Modifiers by Residents of Cairo, Egypt in 2012-2014.\" University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[7]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "expressive2002"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2002"
#> 
#> Slot "components":
#> [1] "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/expressive_acts.htm"
#> 
#> Slot "description":
#> [1] "Appendix of Lisa Slattery Rashotte, 2002, What does that smile mean? The meaning of nonverbal behaviors in social interaction. Social Psychology Quarterly 65: 92-102. EPA values are means across 230 females and 172 males; these cross-sex EPA profiles are given in both the \"male\" and the \"female\" positions."
#> 
#> Slot "citation":
#> [1] "Appendix of Lisa Slattery Rashotte, 2002, What does that smile mean? The meaning of nonverbal behaviors in social interaction. Social Psychology Quarterly 65: 92-102."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[8]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "gaymensanfrancisco1980"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "1980"
#> 
#> Slot "components":
#> [1] "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "unsafebetter" "safebetter"  
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/gay_sex.htm"
#> 
#> Slot "description":
#> [1] "Data were gathered from ten San Francisco gay men in the 1980s by Professor Don Barrett, California State University, San Marcos. All data are from males. The 'male' dictionary present the mean EPA ratings of those with traditional sentiments about unsafe sex practices, feeling that those practices are more pleasurable than new practices. The 'female' dictionary shows the sentiments of those who think that safe-sex practices are more pleasurable."
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] "All respondents are men. \"Male\" and \"female\" sets are coded this way for compatibility with Interact. "
#> 
#> 
#> [[9]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "germany1989"
#> 
#> Slot "context":
#> [1] "Germany"
#> 
#> Slot "year":
#> [1] "1989"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 442 Identities, 295 Behaviors, and 67 Modifiers, selected for back-translatability with the 1978 U.S.A. dictionary were obtained with the Attitude program from 520 Mannheim students. Subjects were matched to the American undergraduate population by proportional inclusion of 12 and 13 grade youths at two German Studenten des Grundstudiums and Gymnasiasten, along with subjects from Mannheim University, which attracts students mainly from the Rhein-Neckar region in former West Germany."
#> 
#> Slot "citation":
#> [1] "Schneider, Andreas. Mean Affective Ratings of 804 Concepts by West German Students  in 1989 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[10]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "germany2007"
#> 
#> Slot "context":
#> [1] "Germany"
#> 
#> Slot "year":
#> [1] "2007"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 376 Identities, 393 Behaviors, 331 Modifiers. (Some of the words are classified as 19 settings in Interact.) Ratings were obtained with Surveyor from 1905 subjects (734 male and 1171 female) from all over Germany. The research was advertised as a 'study of language and emotion' in an extensive recruitment campaign including mailing lists from different universities, weblogs, newspaper reports and radio interviews. Most of the participants (N = 1029) were between 20 and 29 years of age, but the sample covered all ages, including N = 129 being younger than 20 and N = 92 older than 60 years. The data of 83 persons (4.4 %) were excluded from the analysis, as they had indicated German not being their mother tongue."
#> 
#> Slot "citation":
#> [1] "Schroder, Tobias. Mean Affective Ratings of 1100 Concepts by German Adults in 2007 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2013."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[11]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "household1994"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "1994"
#> 
#> Slot "components":
#> [1] "identity" "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/household.htm"
#> 
#> Slot "description":
#> [1] "This set of concepts - developed by Professor Amy Kroska at the University of Oklahoma, and others, for studying gender and household roles - was rated by 23 male and 46 female college students in 1994."
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[12]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "indiana2003"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2003"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 500 Identities, 500 Behaviors, 300 Modifiers, and 200 Settings were collected at Indiana University, via the Internet using the Surveyor applet.  The 1027 respondents lived in the U.S.A. at age 16, and were about equally male and female."
#> 
#> Slot "citation":
#> [1] "Francis, Clare, and David R. Heise. Mean Affective Ratings of 1,500 Concepts by Indiana University Undergraduates in 2002-3 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[13]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "internationaldomesticrelations1981"
#> 
#> Slot "context":
#> [1] "Unknown"
#> 
#> Slot "year":
#> [1] "1981"
#> 
#> Slot "components":
#> [1] "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "professional"    "nonprofessional"
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/international_relations.htm"
#> 
#> Slot "description":
#> [1] "125 inter-state behaviors and 113 domestic behaviors. The behaviors were rated by professionals in international relations, and by individuals from the general population."
#> 
#> Slot "citation":
#> [1] "Azar Edward E. and Steve Lerner 1981. The use of semantic dimensions in the scaling of international events. International Interactions 7: 361-378."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[14]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "internet1998"
#> 
#> Slot "context":
#> [1] "Internet"
#> 
#> Slot "year":
#> [1] "1998"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/InternetCulture.htm"
#> 
#> Slot "description":
#> [1] "This set of concepts related to Internet culture was developed by Adam King, and rated by 2,431 respondents (56%  male and 44% female) in 1998."
#> 
#> Slot "citation":
#> [1] "King, Adam B. 2001. \"Affective dimensions of internet culture.\" Social Science Computer Review 19:414-430."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[15]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "japan1995"
#> 
#> Slot "context":
#> [1] "Japan"
#> 
#> Slot "year":
#> [1] "1989-2002"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 403 Identities and 307 Behaviors, and a few Settings were obtained with the Attitude program from 323 Tohoku University students in 1989. In 1995 and 1996, 120 women students at Kyoritsu Women's, Japan Women's, and Teikyo Universities and 120 men students at Teikyo and Rikkyo Universities rated an additional 300 settings, 300 modifiers (mainly traits), 200 business identities, and 75 behaviors. Yoichi Murase (Rikkyo University) and Nozomu Matsubara (Tokyo University) provided access to students who rated 102 emotions, 70 behaviors and 55 identities in 2002 using Surveyor. Total numbers of entries in Interact lexicon are: 713 Identities, 455 Behaviors, 426 Modifiers, and 300 Settings. Number of male or female raters generally is about 30 for each concept."
#> 
#> Slot "citation":
#> [1] "Smith, Herman W., Takanori Matsuno, Shuuichirou Ike, and Michio Umino. Mean Affective Ratings of 1,894 Concepts by Japanese Undergraduates, 1989-2002 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] "Data collection was in several stages between 1989 and 2002"
#> 
#> 
#> [[16]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "morocco2015"
#> 
#> Slot "context":
#> [1] "Morocco"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///morocco-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Data collected in Morocco. n = 1546, identities = 397, behaviors = 368, modifiers = 233"
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[17]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "nc1978"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "1978"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 721 Identities, 600 Behaviors, 440 Modifiers, and 345 Settings were obtained with paper questionnaires from 1,225 North Carolina undergraduates. (Ratings for some emotion words in this data set were obtained by Heise from Indiana University undergraduates in 1985.) Number of male or female raters generally is about 25 for each word. Funded by National Institute of Mental Health Grant 1-R01-MH29978-01-SSR."
#> 
#> Slot "citation":
#> [1] "Smith-Lovin, Lynn, and David R. Heise. Mean Affective Ratings of 2,106 Concepts by University of North Carolina Undergraduates in 1978 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[18]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "nireland1977"
#> 
#> Slot "context":
#> [1] "North Ireland"
#> 
#> Slot "year":
#> [1] "1977"
#> 
#> Slot "components":
#> [1] "identity" "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 528 Identities and 498 Behaviors were obtained with paper questionnaires from 319 Belfast teenagers in Catholic high schools in 1977. Ratings of modifiers and settings were not obtained in the Belfast study. Up to 18 females and  14 males rated each concept."
#> 
#> Slot "citation":
#> [1] "Willigan, Dennis, and David R. Heise. Mean Affective Ratings of 1,026 Concepts by Catholic High School Students in Belfast, Northern Ireland in 1977 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[19]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "occs2019"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2019"
#> 
#> Slot "components":
#> [1] "identity"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[20]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "occs2020"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2020"
#> 
#> Slot "components":
#> [1] "identity"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[21]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "ontario1980"
#> 
#> Slot "context":
#> [1] "Canada"
#> 
#> Slot "year":
#> [1] "1980-1986"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Data on 843 Identities and  593 Behaviors were obtained from 5,534 Guelph, Ontario, undergraduates with paper questionnaires in 1980-3, and 495 Modifiers rated by 1,260 Guelph undergraduates were added in 1985-6. Funded by the Social Science and Humanities Research Council of Canada."
#> 
#> Slot "citation":
#> [1] "MacKinnon, Neil J. Mean Affective Ratings of 1,931 Concepts by Guelph University Undergraduates, Ontario, Canada, in 1980-6 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[22]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "ontario2001"
#> 
#> Slot "context":
#> [1] "Canada"
#> 
#> Slot "year":
#> [1] "2001-2003"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier" "setting" 
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Data on 993 Identities, 601 Behaviors, and 500 Modifiers, and 200 Settings were gathered with the Attitude program from Guelph, Ontario, undergraduates in 2001-2. Data on settings were gathered with the Surveyor program at Guelph in 2003. Funded by the Social Science and Humanities Research Council of Canada."
#> 
#> Slot "citation":
#> [1] "MacKinnon, Neil J. Mean Affective Ratings of 2,294 Concepts by Guelph University Undergraduates, Ontario, Canada, in 2001-3 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[23]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "politics2003"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2003"
#> 
#> Slot "components":
#> [1] "identity" "behavior"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/interact/subcultures/politics.htm"
#> 
#> Slot "description":
#> [1] "This set of concepts - developed by Kyle Irwin at the University of Missouri-St. Louis - was rated by 47 male and 74 female college students in 2003."
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[24]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "texas1998"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "1998"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean"
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".dat"
#> 
#> Slot "source":
#> [1] "Interact 2.1 beta (May 2021)"
#> 
#> Slot "description":
#> [1] "From https://cs.uwaterloo.ca/~jhoey/research/ACTBackup/ACT/data.html: Ratings of 443 Identities, 278 Behaviors, 65 Modifiers, and 1 Setting were collected at Texas Tech University with the Attitude program. The 482 respondents were nearly equally male and female."
#> 
#> Slot "citation":
#> [1] "Schneider, Andreas. Mean Affective Ratings of 787 Concepts by Texas Tech University Undergraduates in 1998 [Computer file]. Distributed at Affect Control Theory Website, Program Interact <http://www.indiana.edu/~socpsy/ACT/interact/JavaInteract.html>, 2006."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[25]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "uga2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-georgia-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Sentiment dictionary data collected at the University of Georgia. n = 1368, identities = 930, behaviors = 814, modifiers = 660"
#> 
#> Slot "citation":
#> [1] "Robinson, Dawn T., Lynn Smith-Lovin, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. 'Mean Affective Ratings of 932 Identities, 810 Behaviors, and 660 Modifiers by University of Georgia Undergraduates in 2012-2014.'. University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/."
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[26]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "usfullsurveyor2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-combined-surveyor-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Combined sentiment dictionary data collected at the University of Georgia, Duke University, and in the Durham, NC community using Surveyor. n = 1742, identities = 929, behaviors = 814, modifiers = 660"
#> 
#> Slot "citation":
#> [1] "Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. \"Mean Affective Ratings of 929 Identities, 814 Behaviors, and 660 Modifiers by University of Georgia and Duke University Undergraduates and by Community Members in Durham, NC, in 2012-2014.\"\" University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/."
#> 
#> Slot "notes":
#> [1] "A combination of the 2015 Duke student, 2015 UGA student, and 2015 Duke community dictionaries"
#> 
#> 
#> [[27]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "usmturk2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-online-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Collected from Amazon Mechanical Turk workers in 2015. Identities = 968, behaviors = 853, modifiers = 660."
#> 
#> Slot "citation":
#> [1] "Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Brent H. Curdy, and Jonathan H. Morgan. 2019. 'Mean Affective Ratings of 968 Identities, 853 Behaviors, and 660 Modifiers by Amazon Mechanical Turk Workers in 2015.' University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/"
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[28]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "usstudent2015"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2015"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "http://affectcontroltheory.org///usa-student-surveyor-dictionary-2015/"
#> 
#> Slot "description":
#> [1] "Combined sentiment dictionary data collected at the University of Georgia and Duke University using Surveyor. n = 1584, identities = 929, behaviors = 814, modifiers = 660"
#> 
#> Slot "citation":
#> [1] "Smith-Lovin, Lynn, Dawn T. Robinson, Bryan C. Cannon, Jesse K. Clark, Robert Freeland, Jonathan H. Morgan and Kimberly B. Rogers. 2016. 'Mean Affective Ratings of 929 Identities, 814 Behaviors, and 660 Modifiers by University of Georgia and Duke University Undergraduates in 2012-2014.' University of Georgia: Distributed at UGA Affect Control Theory Website: http://research.franklin.uga.edu/act/."
#> 
#> Slot "notes":
#> [1] "A combination of the 2015 Duke student and 2015 UGA student dictionaries"
#> 
#> 
#> [[29]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "artifactmods2022"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2022"
#> 
#> Slot "components":
#> [1] "identity"          "modified_identity" "artifact"         
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".xlsx"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] "Lulham, R., & Shank, D. B. (2022). How “Stuff” Matters in Affect Control Theory. American Behavioral Scientist, 0(0). https://doi.org/10.1177/00027642211066045"
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[30]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "humanvalues2022"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2022"
#> 
#> Slot "components":
#> [1] "value"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[31]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "products2022"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2022"
#> 
#> Slot "components":
#> [1] "artifact"
#> 
#> Slot "stats":
#> [1] "mean" "sd"   "cov" 
#> 
#> Slot "groups":
#> [1] "male"   "female" "all"   
#> 
#> Slot "individual":
#> [1] TRUE
#> 
#> Slot "filetype":
#> [1] ".xlsx"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[32]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "employeeorg2022"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2022"
#> 
#> Slot "components":
#> [1] "identity"
#> 
#> Slot "stats":
#> [1] "mean" "sd"  
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[33]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "ugatech2008"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2008"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[34]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "groups2017"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2017"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[35]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "groups2019"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2019"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[36]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "nounphrasegrammar2019"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2019"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[37]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "generaltech2020"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2020"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[38]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "techvshuman2021"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2021"
#> 
#> Slot "components":
#> [1] ""
#> 
#> Slot "stats":
#> [1] ""
#> 
#> Slot "groups":
#> [1] ""
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] ""
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#> 
#> [[39]]
#> An object of class "dictionary"
#> Slot "key":
#> [1] "mostafaviestimates2022"
#> 
#> Slot "context":
#> [1] "US"
#> 
#> Slot "year":
#> [1] "2022"
#> 
#> Slot "components":
#> [1] "identity" "behavior" "modifier"
#> 
#> Slot "stats":
#> [1] "mean" "sd"  
#> 
#> Slot "groups":
#> [1] "all"
#> 
#> Slot "individual":
#> [1] FALSE
#> 
#> Slot "filetype":
#> [1] ".csv"
#> 
#> Slot "source":
#> [1] "Personal communication"
#> 
#> Slot "description":
#> [1] ""
#> 
#> Slot "citation":
#> [1] ""
#> 
#> Slot "notes":
#> [1] ""
#> 
#>